+++ models/anthropic/CHANGELOG.md
@@ -1,5 +1,38 @@
 # Changelog
 
+## [0.3.8](https://github.com/AIGNE-io/aigne-framework/compare/anthropic-v0.3.7...anthropic-v0.3.8) (2025-06-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.24.1
+  * devDependencies
+    * @aigne/test-utils bumped to 0.4.8
+
+## [0.3.7](https://github.com/AIGNE-io/aigne-framework/compare/anthropic-v0.3.6...anthropic-v0.3.7) (2025-06-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.24.0
+  * devDependencies
+    * @aigne/test-utils bumped to 0.4.7
+
+## [0.3.6](https://github.com/AIGNE-io/aigne-framework/compare/anthropic-v0.3.5...anthropic-v0.3.6) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/core bumped to 1.23.1
+  * devDependencies
+    * @aigne/test-utils bumped to 0.4.6
+
 ## [0.3.5](https://github.com/AIGNE-io/aigne-framework/compare/anthropic-v0.3.4...anthropic-v0.3.5) (2025-06-25)