+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,76 @@
 # Changelog
 
+## [1.14.1](https://github.com/AIGNE-io/aigne-framework/compare/cli-v1.14.0...cli-v1.14.1) (2025-06-26)
+
+
+### Bug Fixes
+
+* agine cli not found package ([#185](https://github.com/AIGNE-io/aigne-framework/issues/185)) ([5d98b61](https://github.com/AIGNE-io/aigne-framework/commit/5d98b6158f1e43e049a3a51a69bab88092bf1c92))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.17.1
+    * @aigne/anthropic bumped to 0.3.8
+    * @aigne/bedrock bumped to 0.3.8
+    * @aigne/core bumped to 1.24.1
+    * @aigne/deepseek bumped to 0.3.8
+    * @aigne/gemini bumped to 0.3.8
+    * @aigne/observability bumped to 0.1.3
+    * @aigne/ollama bumped to 0.3.8
+    * @aigne/open-router bumped to 0.3.8
+    * @aigne/openai bumped to 0.4.1
+    * @aigne/xai bumped to 0.3.8
+
+## [1.14.0](https://github.com/AIGNE-io/aigne-framework/compare/cli-v1.13.2...cli-v1.14.0) (2025-06-26)
+
+
+### Features
+
+* **transport:** support invoke server side chat model ([#182](https://github.com/AIGNE-io/aigne-framework/issues/182)) ([f81a1bf](https://github.com/AIGNE-io/aigne-framework/commit/f81a1bf883abda1845ccee09b270e5f583e287ab))
+
+
+### Bug Fixes
+
+* blocklet start failed ([#180](https://github.com/AIGNE-io/aigne-framework/issues/180)) ([296a481](https://github.com/AIGNE-io/aigne-framework/commit/296a481be69d9b9b279dc4e50b0d21c993d1d841))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.17.0
+    * @aigne/anthropic bumped to 0.3.7
+    * @aigne/bedrock bumped to 0.3.7
+    * @aigne/core bumped to 1.24.0
+    * @aigne/deepseek bumped to 0.3.7
+    * @aigne/gemini bumped to 0.3.7
+    * @aigne/observability bumped to 0.1.2
+    * @aigne/ollama bumped to 0.3.7
+    * @aigne/open-router bumped to 0.3.7
+    * @aigne/openai bumped to 0.4.0
+    * @aigne/xai bumped to 0.3.7
+
+## [1.13.2](https://github.com/AIGNE-io/aigne-framework/compare/cli-v1.13.1...cli-v1.13.2) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.16.1
+    * @aigne/anthropic bumped to 0.3.6
+    * @aigne/bedrock bumped to 0.3.6
+    * @aigne/core bumped to 1.23.1
+    * @aigne/deepseek bumped to 0.3.6
+    * @aigne/gemini bumped to 0.3.6
+    * @aigne/ollama bumped to 0.3.6
+    * @aigne/open-router bumped to 0.3.6
+    * @aigne/openai bumped to 0.3.6
+    * @aigne/xai bumped to 0.3.6
+
 ## [1.13.1](https://github.com/AIGNE-io/aigne-framework/compare/cli-v1.13.0...cli-v1.13.1) (2025-06-25)